requirements.txt
langchain
langchain-openai
fastapi
pydantic
uvicorn
python-dotenv